Speed up wrapper script when no edl provided
Hugh correctly noted in #185 that the script doesn't need to do a merge
if no edl file is given.  This fixes that.

Also took this opportunity to clean up the script a bit...
  - the call to ffmpeg has been moved into the function which now not
  only monitors ffmpeg but also launches it and checks the output,
  renamed to "launch_and_monitor_ffmpeg"
  - reduced the use of global variables, instead passing by argument and
  using the bash "local" keyword.

+++ cTiVo/cTiVoScripts/ffmpeg_edl_ac3.sh
@@ -15,57 +15,78 @@ exit 1
 }
 
 timestamp_to_seconds() {
-  hours=$(echo $1 | cut -d: -f1)
-  minutes=$(echo $1 | cut -d: -f2)
-  seconds=$(echo $1 | cut -d: -f3)
+  local timestamp="$1"
+  hours=$(echo $timestamp | cut -d: -f1)
+  minutes=$(echo $timestamp | cut -d: -f2)
+  seconds=$(echo $timestamp | cut -d: -f3)
   seconds_fraction=$(echo "$seconds" | cut -d, -f2 | awk '{printf("%03d",$1)}')
   seconds=$(echo "$seconds" | cut -d, -f1)
   echo $(echo "3600*$hours + 60*$minutes + $seconds + $seconds_fraction/1000" | bc -l)
 }
 
-# monitor the ffmpeg proc as it runs in the background, grepping the time out
+# launch and monitor the ffmpeg proc as it runs in the background, grepping the time out
 # of the end of the ffmpeg log for the progress indicator
-monitor() {
-# Globals:
-#  $pid: process for ffmpeg
-#  $ffmpeg_logfile: path to log file
-#  $this_duration: expected length of current segment of video in seconds
-#  $min_percent -> $max_percent: the range into which we map ffmpeg progress for reporting (0-100)
-
+# Arguments:
+#  min_percent:
+#  max_percent: the range into which we map ffmpeg progress for reporting (0-100)
+#  duration: expected length of current segment of video in seconds
+#  logfile: path to log file
+#
 # cTiVo will kill this script if it continuously outputs the same progress % for a
 # MaxProgressDelay (default 2 min) window.
+launch_and_monitor_ffmpeg() {
+  local min_percent="$1"; shift
+  local max_percent="$1"; shift
+  local duration="$1"; shift
+  local output="$1"; shift
+  local logfile="$1"; shift
+  local progress percent last_percent
+  set -x
+  "$ffmpeg_path" "$@" "$output" >& "$logfile" &
+  pid=$!
+  set +x
+  last_percent="$min_percent"
+  echo "$last_percent %"
   while kill -0 $pid &> /dev/null; do
-    this_progress=$(tail -c 1000 "$ffmpeg_logfile"  | egrep -o 'time=\S+' | cut -d= -f2 | tail -n1)
-    if [ ! -z "$this_progress" ]; then
-      this_progress=$(timestamp_to_seconds "$this_progress")
-      pct=$(echo "$min_percent $max_percent $this_progress $this_duration " | awk '{printf("%.2f", $1 + ($2 - $1) * $3 / $4)}')
-      if (($(echo "$pct <  $min_percent" | bc -l))); then pct=$min_percent; fi
-      if (($(echo "$pct >= $max_percent" | bc -l))); then pct=$max_percent; fi
-      if [ "$last_pct" != "$pct" ]; then
-        echo "$pct %"
+    progress=$(tail -c 1000 "$logfile"  | egrep -o 'time=\S+' | cut -d= -f2 | tail -n1)
+    if [ -n "$progress" ]; then
+      progress=$(timestamp_to_seconds "$progress")
+      percent=$(echo "$min_percent $max_percent $progress $duration " | awk '{printf("%.2f", $1 + ($2 - $1) * $3 / $4)}')
+      if [ "$last_percent" != "$percent" ]; then
+        echo "$percent %"
       fi
-      last_pct="$pct"
+      last_percent="$percent"
     fi
     sleep 1
   done 
+  last_percent="$max_percent"
+  echo "$last_percent %"
+  pid=
+
+  if [ ! -f "$output" ]; then
+    echo "error: problem generating $output, see $logfile for details"
+    exit 1
+  fi
 }
 
 # kill any encode that's going on in the background before 
 # exiting ourselves
 terminate() {
-  kill -TERM "$pid" 2> /dev/null
+  if [ -n $pid ]; then
+    kill -TERM "$pid" 2> /dev/null
+  fi
   echo "Terminating"
   exit 15
 }
 
 while (( $# > 0 )); do
-  if [ "$1" == "-h" ]; then usage; fi
-  if [ "$1" == "-i" ]; then
+  if [ "$1" == "-h" ]; then
+    usage
+  elif [ "$1" == "-i" ]; then
     input="$2"
     shift 2
     continue
-  fi
-  if [ "$1" == "-edl" ]; then
+  elif [ "$1" == "-edl" ]; then
     edl_file="$2"
     shift 2
     continue
@@ -85,48 +106,24 @@ while (( $# > 0 )); do
 done
 
 if [ -z "$output" ]; then
-  echo "missing output"
+  echo "missing output file"
   usage
-fi
-if [ -z "$input" ]; then
+elif [ -z "$input" ]; then
   echo "missing -i"
   usage
-fi
-
-if [ ! -f "$input" ]; then
+elif [ ! -f "$input" ]; then
   echo "no such file: $input"
   usage
 fi
-
-awkcmd='
-BEGIN {OFS=":"}
-{
-  if (NR == 1 && $1 > 0) {
-    print "",$1
-  } else {
-    if (NR > 1) {
-      print ss,$1
-    }
-  }
-  ss = $2
-}
-END {print ss,""}
-'
+ext="${output##*.}"
 
 tmpdir="${input%.*}_ffmpeg"
 rm -rf "$tmpdir"
 mkdir -p "$tmpdir/comskip"
 mkdir -p "$tmpdir/logs"
 
-if [ -z "$edl_file" ]; then
-  edl_file="$tmpdir/dummy.edl"
-  touch "$edl_file"
-fi
-
-ext="${output##*.}"
-
 # figure out the duration after cutting, for the progress indicator
-file_info=$($ffmpeg_path -i "$input" 2>&1 >/dev/null)
+file_info=$("$ffmpeg_path" -i "$input" 2>&1 >/dev/null)
 original_duration=$(echo "$file_info" | grep Duration: | awk '{print $2}')
 original_duration=$(timestamp_to_seconds $original_duration)
 cut_duration=$(awk "BEGIN {total=0} {total += ($original_duration < \$2 ? $original_duration : \$2)-\$1} END {print total}" "$edl_file")
@@ -144,65 +141,63 @@ if echo "$file_info" | grep -m1 Audio: | grep ac3 | grep --quiet '5\.1'; then
   ac3_opts="-c:a:1 ac3"
 fi
 
-ffmpeg_concat_filename="$tmpdir/ffmpeg_concat.txt"
-i=1
-progress=0
-for startstop in $(/usr/bin/awk "$awkcmd" "$edl_file"); do
-  ss=$(echo $startstop | cut -d: -f1)
-  to=$(echo $startstop | cut -d: -f2)
+if [ -z "$edl_file" ]; then
+  # no edl file, don't need to encode segments and merge, simply encode with the modified audio streams
+  args=($ffmpeg_opts_pre_input -i "$input" $map_opts $ffmpeg_opts_post_input -strict -2 -c:a:0 aac $ac3_opts)
+  launch_and_monitor_ffmpeg 0 100 $duration "$output" "$tmpdir/ffmpeg.txt"
 
-  # "input seeking" (-ss before the input file) in ffmpeg is very fast but
-  # the timestamps are reset so -to is now really a duration rather than a timestamp to stop at
-  if [ ! -z "$to" ]; then
-    if [ ! -z "$ss" ]; then to=$(echo "$to - $ss" | bc -l | awk '{printf("%.3f",$1)}'); fi
-    this_duration="$to"
-    to="-to $to"
-  else
-    if [ -z "$ss" ]; then ss="0"; fi
-    this_duration=$(echo "$original_duration - $ss" | bc -l)
-  fi
-  if [ ! -z "$ss" ]; then
-    # skip if the beginning of the cut is longer than the duration of the video
-    if (($(echo "$ss > $original_duration" | bc -l))); then continue; fi
-    ss="-ss $ss";
-  fi
+else
+  # encode segments separately and merge together
+  i=1
+  progress=0
+  for startstop in $(/usr/bin/awk '
+  BEGIN {OFS=":"}
+  {
+    if (NR == 1 && $1 > 0) {
+      print "",$1
+    } else {
+      if (NR > 1) {
+        print ss,$1
+      }
+    }
+    ss = $2
+  }
+  END {print ss,""}
+  ' "$edl_file"); do
+    ss=$(echo $startstop | cut -d: -f1)
+    to=$(echo $startstop | cut -d: -f2)
 
-  segment_filename="segment${i}.$ext"
-  ffmpeg_logfile="$tmpdir/logs/ffmpeg_segment$i.log"
-  set -x
-  $ffmpeg_path $ss $ffmpeg_opts_pre_input -i "$input" $map_opts $ffmpeg_opts_post_input -strict -2 -c:a:0 aac $ac3_opts $to "$tmpdir/$segment_filename" >& "$ffmpeg_logfile" &
-  pid=$!
-  set +x
+    # "input seeking" (-ss before the input file) in ffmpeg is very fast but
+    # the timestamps are reset so -to is now really a duration rather than a timestamp to stop at
+    if [[ -n "$to" && -n "$ss" ]]; then
+      this_duration=$(echo "$to - $ss" | bc -l | /usr/bin/awk '{printf("%.3f",$1)}')
+    elif [[ -n "$ss" ]]; then
+      if (($(echo "$ss > $original_duration" | bc -l))); then
+        continue
+      fi
+      this_duration=$(echo "$original_duration - $ss" | bc -l)
+    fi
+    if [[ -n "$ss" ]]; then
+      ss="-ss $ss"
+    fi
+    if [[ -n "$to" ]]; then
+      to="-to $this_duration"
+    fi
 
-  # monitor ffmpeg and update progress
-  min_percent=$(echo "90 * ($progress / $duration)" | bc -l)
-  max_percent=$(echo "90 * (($progress + $this_duration) / $duration)" | bc -l)
-  monitor
+    launch_and_monitor_ffmpeg $(echo "90 * ($progress / $duration)" | bc -l) \
+                              $(echo "90 * (($progress + $this_duration) / $duration)" | bc -l) \
+                              $this_duration \
+                              "$tmpdir/segment${i}.$ext" \
+                              "$tmpdir/logs/ffmpeg_segment$i.log" \
+                              $ss $ffmpeg_opts_pre_input -i "$input" $map_opts $ffmpeg_opts_post_input -strict -2 -c:a:0 aac $ac3_opts $to "$segment_filename"
 
-  if [ -f "$tmpdir/$segment_filename" ]; then
-    echo "file $segment_filename" >> "$ffmpeg_concat_filename"
-  else
-    echo "error: problem generating $segment_filename, see $ffmpeg_logfile for details"
-    exit 1
-  fi
-  let i++
-  progress=$(echo "$this_duration + $progress" | bc -l)
-done
+    progress=$(echo "$this_duration + $progress" | bc -l)
+    let i++
+  done
 
-# merge the segments together
-# though this usually happens in under 2 minutes, it could 
-# take longer for larger videos, so restart progress at 0 
-# and track progress for the merge
-rm -f "$output"
-ffmpeg_logfile="$tmpdir/logs/ffmpeg_concat.log"
-set -x
-$ffmpeg_path -f concat -safe 0 -i "$ffmpeg_concat_filename" $map_opts -c copy $ac3_opts "$output" >& "$ffmpeg_logfile" &
-pid=$!
-set +x
-
-this_duration=$duration
-min_percent=90
-max_percent=100
-monitor
-
-echo "100.00 %"
+  # merge the segments together
+  # though this usually happens in under 2 minutes, it could 
+  # take longer for larger videos, so track progress for the merge
+  rm -f "$output"
+  launch_and_monitor_ffmpeg 90 100 $duration "$tmpdir/logs/ffmpeg_concat.log" -f concat -safe 0 -i <"$(ls $tmpdir/segment*.$ext)" $map_opts -c copy $ac3_opts "$output"
+endif
